requirements.txt
django==4.0.5
django-phonenumber-field==6.3.0
environs==9.5.0
phonenumbers==8.12.50
python-telegram-bot==13.13
python-dateutil==2.8.2